Use line-by-line over hunks in `lazygit`
It seems that as per the latest release of `lazygit` the `hunk` method
for staging the changes on a file is by hunks rather than the former
line-by-line which I'm more comfortable with. Alternatively, given that
it can be easily switched with `a`, this is a matter of using
line-by-line as the default instead.